requirements.txt
matplotlib==3.7.1
numpy==1.24.2
pandas==1.5.3
pandas-datareader==0.10.0
Pillow==9.4.0
scikit-learn==1.2.2